+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [@adguard/logger] v2.0.0 for logging.
+
 ## [v2.0.3] - 2025-06-26
 
-## Added
+### Added
 
 - [@adguard/agtree] v3.2.1 for rules conversion and validation.
 - [@adguard/ecss-tree] v1.1.0 for selector lists validation.